+++ rmf_traffic_ros2/src/rmf_traffic_ros2/schedule/internal_Node.hpp
@@ -53,6 +53,8 @@
 
 #include <rmf_traffic_msgs/msg/negotiation_notice.hpp>
 
+#include <rmf_traffic_ros2/schedule/ParticipantRegistry.hpp>
+
 #include <rmf_utils/Modular.hpp>
 
 #include <set>
@@ -401,10 +403,12 @@ public:
     std::unordered_map<ParticipantId, Wait> _waiting;
     std::shared_ptr<const rmf_traffic::schedule::Snappable> _viewer;
     Version _next_negotiation_version = 0;
+    
   };
 
   ConflictRecord active_conflicts;
   std::mutex active_conflicts_mutex;
+  std::shared_ptr<ParticipantRegistry> participant_registry;
 };
 
 } // namespace schedule