+++ rmf_traffic_ros2/CHANGELOG.rst
@@ -2,6 +2,10 @@
 Changelog for package rmf_traffic_ros2
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+1.3.0 (2021-XX-XX)
+------------------
+* Add persistence to Traffic Schedule Participant IDs: [#242](https://github.com/osrf/rmf_core/pull/242)
+
 1.2.0 (2021-01-05)
 ------------------
 * Adding distributed blockade system hooks: [#226](https://github.com/osrf/rmf_core/pull/226)